Move back interfaces to proper place to respect backward compatibility (#1336)
* 2.8.x breaks this, and we're about to rollback
